+++ api_app/db/repositories/shared_services.py
@@ -6,11 +6,9 @@ from pydantic import parse_obj_as
 from db.repositories.resource_templates import ResourceTemplateRepository
 
 from db.repositories.resources import ResourceRepository, IS_ACTIVE_CLAUSE
-from db.repositories.operations import OperationRepository
 from models.domain.shared_service import SharedService
 from models.schemas.resource import ResourcePatch
-from models.schemas.shared_service import SharedServiceInCreate
-from db.errors import ResourceIsNotDeployed, EntityDoesNotExist
+from models.schemas.shared_service_template import SharedServiceInCreate
 from models.domain.resource import ResourceType
 
 
@@ -24,7 +22,7 @@ class SharedServiceRepository(ResourceRepository):
 
     @staticmethod
     def active_shared_services_query(shared_service_id: str):
-        return f'SELECT * FROM c WHERE {IS_ACTIVE_CLAUSE} AND c.resourceType = "{ResourceType.SharedService}" AND c.sharedSerivceId = "{shared_service_id}"'
+        return f'SELECT * FROM c WHERE {IS_ACTIVE_CLAUSE} AND c.resourceType = "{ResourceType.SharedService}" AND c.sharedId = "{shared_service_id}"'
 
     def get_active_shared_services_for_shared(self, shared_service_id: str) -> List[SharedService]:
         """
@@ -33,21 +31,6 @@ class SharedServiceRepository(ResourceRepository):
         query = SharedServiceRepository.active_shared_services_query(shared_service_id)
         shared_services = self.query(query=query)
         return parse_obj_as(List[SharedService], shared_services)
-
-    def get_deployed_shared_service_by_id(self, shared_service_id: str, service_id: str, operations_repo: OperationRepository) -> SharedService:
-        shared_service = self.get_shared_service_by_id(shared_service_id, service_id)
-
-        if (not operations_repo.resource_has_deployed_operation(resource_id=service_id)):
-            raise ResourceIsNotDeployed
-
-        return shared_service
-
-    def get_shared_service_by_id(self, shared_service_id: str, service_id: str) -> SharedService:
-        query = self.shared_services_query(shared_service_id) + f' AND c.id = "{service_id}"'
-        shared_services = self.query(query=query)
-        if not shared_services:
-            raise EntityDoesNotExist
-        return parse_obj_as(SharedService, shared_services[0])
 
     def get_shared_service_spec_params(self):
         return self.get_resource_base_spec_params()
@@ -61,8 +44,7 @@ class SharedServiceRepository(ResourceRepository):
         resource_spec_parameters = {**shared_service_input.properties, **self.get_shared_service_spec_params()}
 
         shared_service = SharedService(
-            id=full_shared_service_id,
-            sharedServiceId=shared_service_id,
+            sharedServiceIid=full_shared_service_id,
             templateName=shared_service_input.templateName,
             templateVersion=template_version,
             properties=resource_spec_parameters,
