+++ api_app/models/domain/shared_service.py
@@ -1,0 +1,8 @@
+from models.domain.resource import Resource, ResourceType
+
+
+class SharedService(Resource):
+    """
+    Shared service request
+    """
+    resourceType = ResourceType.WorkspaceService
