+++ api_app/tests_ma/conftest.py
@@ -5,6 +5,7 @@ from models.domain.user_resource_template import UserResourceTemplate
 from models.schemas.user_resource_template import UserResourceTemplateInCreate, UserResourceTemplateInResponse
 from models.schemas.workspace_template import WorkspaceTemplateInCreate
 from models.schemas.workspace_service_template import WorkspaceServiceTemplateInCreate
+from models.schemas.shared_service_template import SharedServiceTemplateInCreate
 
 
 @pytest.fixture
@@ -75,6 +76,23 @@ def input_user_resource_template():
             }
         ])
 
+
+@pytest.fixture
+def input_shared_service_template():
+    return SharedServiceTemplateInCreate(
+        name="my-tre-shared-service",
+        version="0.0.1",
+        current=True,
+        json_schema={
+            "$schema": "http://json-schema.org/draft-07/schema",
+            "$id": "https://github.com/microsoft/AzureTRE/templates/shared-services/shared-service.json",
+            "type": "object",
+            "title": "My Shared Service Template",
+            "description": "This is a test shared service template schema.",
+            "required": [],
+            "properties": {}
+        }
+    )
 
 @pytest.fixture
 def basic_resource_template(input_workspace_template):
