+++ api_app/tests_ma/test_db/test_repositories/test_resource_templates_repository.py
@@ -2,10 +2,9 @@ import pytest
 from mock import patch
 
 from db.repositories.resource_templates import ResourceTemplateRepository
-from db.errors import DuplicateEntity, EntityDoesNotExist
+from db.errors import EntityDoesNotExist
 from models.domain.resource import ResourceType
 from models.domain.resource_template import ResourceTemplate
-from models.domain.user_resource_template import UserResourceTemplate
 
 
 @pytest.fixture
@@ -14,7 +13,7 @@ def resource_template_repo():
         yield ResourceTemplateRepository(cosmos_client_mock)
 
 
-def sample_resource_template(name: str, version: str = "1.0", resource_type: ResourceType = ResourceType.Workspace) -> ResourceTemplate:
+def sample_resource_template_as_dict(name: str, version: str = "1.0", resource_type: ResourceType = ResourceType.Workspace) -> dict:
     return ResourceTemplate(
         id="a7a7a7bd-7f4e-4a4e-b970-dc86a6b31dfb",
         name=name,
@@ -25,16 +24,7 @@ def sample_resource_template(name: str, version: str = "1.0", resource_type: Res
         properties={},
         actions=[],
         required=[]
-    )
-
-
-def sample_resource_template_as_dict(name: str, version: str = "1.0", resource_type: ResourceType = ResourceType.Workspace) -> dict:
-    return sample_resource_template(name, version, resource_type).dict()
-
-
-def sample_user_resource_template_as_dict(name: str, version: str = "1.0") -> dict:
-    template = UserResourceTemplate(id="123", name=name, description="", version=version, current=True, required=[], properties={}, actions=[], parentWorkspaceService="parent_service")
-    return template.dict()
+    ).dict()
 
 
 @patch('db.repositories.resource_templates.ResourceTemplateRepository.query')
@@ -43,16 +33,6 @@ def test_get_by_name_and_version_queries_db(query_mock, resource_template_repo):
     query_mock.return_value = [sample_resource_template_as_dict(name="test", version="1.0")]
 
     resource_template_repo.get_template_by_name_and_version(name="test", version="1.0", resource_type=ResourceType.Workspace)
-
-    query_mock.assert_called_once_with(query=expected_query)
-
-
-@patch('db.repositories.resource_templates.ResourceTemplateRepository.query')
-def test_get_user_resource_template_by_name_and_version_queries_db(query_mock, resource_template_repo):
-    expected_query = 'SELECT * FROM c WHERE c.resourceType = "user-resource" AND c.name = "test" AND c.version = "1.0" AND c.parentWorkspaceService = "parent_service"'
-    query_mock.return_value = [sample_user_resource_template_as_dict(name="test", version="1.0")]
-
-    resource_template_repo.get_template_by_name_and_version(name="test", version="1.0", resource_type=ResourceType.UserResource, parent_service_name="parent_service")
 
     query_mock.assert_called_once_with(query=expected_query)
 
@@ -109,47 +89,6 @@ def test_get_current_by_name_raises_entity_does_not_exist_if_no_template_found(q
 
 
 @patch('db.repositories.resource_templates.ResourceTemplateRepository.query')
-def test_get_current_user_resource_template_queries_db(query_mock, resource_template_repo):
-    template_name = "template1"
-    parent_template_name = "parent_template1"
-    expected_query = 'SELECT * FROM c WHERE c.resourceType = "user-resource" AND c.name = "template1" AND c.current = true AND c.parentWorkspaceService = "parent_template1"'
-    query_mock.return_value = [sample_resource_template_as_dict(name=template_name)]
-
-    resource_template_repo.get_current_template(template_name, ResourceType.UserResource, parent_template_name)
-
-    query_mock.assert_called_once_with(query=expected_query)
-
-
-@patch('db.repositories.resource_templates.ResourceTemplateRepository.query')
-def test_get_current_user_resource_template_returns_matching_template(query_mock, resource_template_repo):
-    template_name = "template1"
-    parent_template_name = "parent_template1"
-    query_mock.return_value = [sample_resource_template_as_dict(name=template_name)]
-
-    template = resource_template_repo.get_current_template(template_name, ResourceType.UserResource, parent_template_name)
-
-    assert template.name == template_name
-
-
-@patch('db.repositories.resource_templates.ResourceTemplateRepository.query')
-def test_get_current_user_resource_template_raises_entity_does_not_exist_if_no_template_found(query_mock, resource_template_repo):
-    query_mock.return_value = []
-
-    with pytest.raises(EntityDoesNotExist):
-        resource_template_repo.get_current_template("template1", ResourceType.UserResource, "parent_template1")
-
-
-@patch('db.repositories.resource_templates.ResourceTemplateRepository.query')
-def test_get_current_user_resource_template_raises_duplicate_entity_if_multiple_current_found(query_mock, resource_template_repo):
-    template_name = "template1"
-    parent_template_name = "parent_template1"
-    query_mock.return_value = [sample_resource_template_as_dict(name=template_name), sample_resource_template_as_dict(name=template_name)]
-
-    with pytest.raises(DuplicateEntity):
-        resource_template_repo.get_current_template(template_name, ResourceType.UserResource, parent_template_name)
-
-
-@patch('db.repositories.resource_templates.ResourceTemplateRepository.query')
 def test_get_templates_information_returns_unique_template_names(query_mock, resource_template_repo):
     query_mock.return_value = [
         {"name": "template1", "title": "title1", "description": "description1"},
@@ -165,54 +104,7 @@ def test_get_templates_information_returns_unique_template_names(query_mock, res
 
 @patch('db.repositories.resource_templates.ResourceTemplateRepository.save_item')
 @patch('uuid.uuid4')
-def test_create_workspace_template_item_calls_create_item_with_the_correct_parameters(uuid_mock, save_item_mock, resource_template_repo, input_workspace_template):
-    uuid_mock.return_value = "1234"
-
-    returned_template = resource_template_repo.create_template(input_workspace_template, ResourceType.Workspace)
-
-    expected_resource_template = ResourceTemplate(
-        id="1234",
-        name=input_workspace_template.name,
-        title=input_workspace_template.json_schema["title"],
-        description=input_workspace_template.json_schema["description"],
-        version=input_workspace_template.version,
-        resourceType=ResourceType.Workspace,
-        properties=input_workspace_template.json_schema["properties"],
-        actions=input_workspace_template.customActions,
-        required=input_workspace_template.json_schema["required"],
-        current=input_workspace_template.current
-    )
-    save_item_mock.assert_called_once_with(expected_resource_template)
-    assert expected_resource_template == returned_template
-
-
-@patch('db.repositories.resource_templates.ResourceTemplateRepository.save_item')
-@patch('uuid.uuid4')
-def test_create_user_resource_template_item_calls_create_item_with_the_correct_parameters(uuid_mock, save_item_mock, resource_template_repo, input_user_resource_template):
-    uuid_mock.return_value = "1234"
-
-    returned_template = resource_template_repo.create_template(input_user_resource_template, ResourceType.UserResource, "parent_service_template_name")
-
-    expected_resource_template = UserResourceTemplate(
-        id="1234",
-        name=input_user_resource_template.name,
-        title=input_user_resource_template.json_schema["title"],
-        description=input_user_resource_template.json_schema["description"],
-        version=input_user_resource_template.version,
-        resourceType=ResourceType.UserResource,
-        properties=input_user_resource_template.json_schema["properties"],
-        actions=input_user_resource_template.customActions,
-        required=input_user_resource_template.json_schema["required"],
-        current=input_user_resource_template.current,
-        parentWorkspaceService="parent_service_template_name"
-    )
-    save_item_mock.assert_called_once_with(expected_resource_template)
-    assert expected_resource_template == returned_template
-
-
-@patch('db.repositories.resource_templates.ResourceTemplateRepository.save_item')
-@patch('uuid.uuid4')
-def test_create_item_created_with_the_expected_type(uuid_mock, save_item_mock, resource_template_repo, input_workspace_template):
+def test_create_item_created_with_the_correct_parameters_and_type(uuid_mock, save_item_mock, resource_template_repo, input_workspace_template):
     uuid_mock.return_value = "1234"
     expected_type = ResourceType.WorkspaceService
     returned_template = resource_template_repo.create_template(input_workspace_template, expected_type)
@@ -230,13 +122,3 @@ def test_create_item_created_with_the_expected_type(uuid_mock, save_item_mock, r
     )
     save_item_mock.assert_called_once_with(expected_resource_template)
     assert expected_resource_template == returned_template
-
-
-@patch('db.repositories.resource_templates.ResourceTemplateRepository.query')
-def test_get_template_infos_for_user_resources_queries_db(query_mock, resource_template_repo):
-    expected_query = 'SELECT c.name, c.title, c.description FROM c WHERE c.resourceType = "user-resource" AND c.current = true AND c.parentWorkspaceService = "parent_service"'
-    query_mock.return_value = [sample_resource_template_as_dict(name="test", version="1.0", resource_type=ResourceType.UserResource)]
-
-    resource_template_repo.get_templates_information(ResourceType.UserResource, parent_service_name="parent_service")
-
-    query_mock.assert_called_once_with(query=expected_query)
